requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
python-dotenv==1.0.0
